+++ federatedscope/core/aggregators/__init__.py
@@ -1,0 +1,19 @@
+from federatedscope.core.aggregators.aggregator import Aggregator, \
+    NoCommunicationAggregator
+from federatedscope.core.aggregators.clients_avg_aggregator import \
+    ClientsAvgAggregator, OnlineClientsAvgAggregator
+from federatedscope.core.aggregators.asyn_clients_avg_aggregator import \
+    AsynClientsAvgAggregator
+from federatedscope.core.aggregators.server_clients_interpolate_aggregator \
+    import ServerClientsInterpolateAggregator
+from federatedscope.core.aggregators.fedopt_aggregator import FedOptAggregator
+
+__all__ = [
+    'Aggregator',
+    'NoCommunicationAggregator',
+    'ClientsAvgAggregator',
+    'OnlineClientsAvgAggregator',
+    'AsynClientsAvgAggregator',
+    'ServerClientsInterpolateAggregator',
+    'FedOptAggregator',
+]
